adpd410x: fix code style

+++ iio_adpd410x.c
@@ -61,9 +61,9 @@
  * @param channel - IIO channel information.
  * @return Number of bytes printed in the output buffer, or negative error code.
  */
-static ssize_t adpd410x_iio_read_raw_chan(void *device, char *buf,
-		size_t len,
-		const struct iio_ch_info *channel, intptr_t priv)
+static int adpd410x_iio_read_raw_chan(void *device, char *buf, uint32_t len,
+				      const struct iio_ch_info *channel,
+				      intptr_t priv)
 {
 	struct adpd410x_dev *dev = (struct adpd410x_dev *)device;
 	int32_t ret;
@@ -92,8 +92,8 @@ static ssize_t adpd410x_iio_read_raw_chan(void *device, char *buf,
  * @param channel - IIO channel information.
  * @return Number of bytes printed in the output buffer, or negative error code.
  */
-static ssize_t adpd410x_iio_set_sampling_freq(void *device, char *buf,
-		size_t len,
+static int adpd410x_iio_set_sampling_freq(void *device, char *buf,
+		uint32_t len,
 		const struct iio_ch_info *channel, intptr_t priv)
 {
 	struct adpd410x_dev *dev = (struct adpd410x_dev *)device;
@@ -116,8 +116,8 @@ static ssize_t adpd410x_iio_set_sampling_freq(void *device, char *buf,
  * @param channel - IIO channel information.
  * @return Number of bytes printed in the output buffer, or negative error code.
  */
-static ssize_t adpd410x_iio_get_sampling_freq(void *device, char *buf,
-		size_t len,
+static int adpd410x_iio_get_sampling_freq(void *device, char *buf,
+		uint32_t len,
 		const struct iio_ch_info *channel, intptr_t priv)
 {
 	struct adpd410x_dev *dev = (struct adpd410x_dev *)device;
@@ -143,8 +143,8 @@ static char const * const adpd410x_iio_timeslots[] = {
  * @param channel - IIO channel information.
  * @return Number of bytes printed in the output buffer, or negative error code.
  */
-static ssize_t adpd410x_iio_set_last_timeslot(void *device, char *buf,
-		size_t len,
+static int adpd410x_iio_set_last_timeslot(void *device, char *buf,
+		uint32_t len,
 		const struct iio_ch_info *channel, intptr_t priv)
 {
 	struct adpd410x_dev *dev = (struct adpd410x_dev *)device;
@@ -170,8 +170,8 @@ static ssize_t adpd410x_iio_set_last_timeslot(void *device, char *buf,
  * @param channel - IIO channel information.
  * @return Number of bytes printed in the output buffer, or negative error code.
  */
-static ssize_t adpd410x_iio_get_last_timeslot(void *device, char *buf,
-		size_t len,
+static int adpd410x_iio_get_last_timeslot(void *device, char *buf,
+		uint32_t len,
 		const struct iio_ch_info *channel, intptr_t priv)
 {
 	struct adpd410x_dev *dev = (struct adpd410x_dev *)device;
@@ -193,8 +193,8 @@ static ssize_t adpd410x_iio_get_last_timeslot(void *device, char *buf,
  * @param channel - IIO channel information.
  * @return Number of bytes printed in the output buffer, or negative error code.
  */
-static ssize_t adpd410x_iio_get_last_timeslot_available(void *device, char *buf,
-		size_t len,
+static int adpd410x_iio_get_last_timeslot_available(void *device, char *buf,
+		uint32_t len,
 		const struct iio_ch_info *channel, intptr_t priv)
 {
 	strcpy(buf, "");
@@ -219,8 +219,9 @@ static char const * const adpd410x_iio_opmode[] = {
  * @param channel - IIO channel information.
  * @return Number of bytes printed in the output buffer, or negative error code.
  */
-static ssize_t adpd410x_iio_set_opmode(void *device, char *buf, size_t len,
-				       const struct iio_ch_info *channel, intptr_t priv)
+static int adpd410x_iio_set_opmode(void *device, char *buf, uint32_t len,
+				   const struct iio_ch_info *channel,
+				   intptr_t priv)
 {
 	struct adpd410x_dev *dev = (struct adpd410x_dev *)device;
 	int32_t ret;
@@ -245,8 +246,9 @@ static ssize_t adpd410x_iio_set_opmode(void *device, char *buf, size_t len,
  * @param channel - IIO channel information.
  * @return Number of bytes printed in the output buffer, or negative error code.
  */
-static ssize_t adpd410x_iio_get_opmode(void *device, char *buf, size_t len,
-				       const struct iio_ch_info *channel, intptr_t priv)
+static int adpd410x_iio_get_opmode(void *device, char *buf, uint32_t len,
+				   const struct iio_ch_info *channel,
+				   intptr_t priv)
 {
 	struct adpd410x_dev *dev = (struct adpd410x_dev *)device;
 	int32_t ret;
@@ -267,8 +269,8 @@ static ssize_t adpd410x_iio_get_opmode(void *device, char *buf, size_t len,
  * @param channel - IIO channel information.
  * @return Number of bytes printed in the output buffer, or negative error code.
  */
-static ssize_t adpd410x_iio_get_opmode_available(void *device, char *buf,
-		size_t len,
+static int adpd410x_iio_get_opmode_available(void *device, char *buf,
+		uint32_t len,
 		const struct iio_ch_info *channel, intptr_t priv)
 {
 	strcpy(buf, "");
